Add free eBook: The Road to learn React
Okay, my second and last attempt to add the book :) The book is perceived as an open source book to learn React. People like and share it in the community. I would love to see it in the list of React resources. 

Even though people can pay what they want to support events like this: https://www.robinwieruch.de/giving-back-by-learning-react/